Moved defined mode to new file
